+++ app/app.go
@@ -16,13 +16,13 @@ import (
 	"github.com/madflojo/hord"
 	"github.com/madflojo/hord/drivers/cassandra"
 	"github.com/madflojo/hord/drivers/redis"
-	"github.com/madflojo/tarmac/callbacks"
-	"github.com/madflojo/tarmac/callbacks/httpclient"
-	"github.com/madflojo/tarmac/callbacks/kvstore"
-	"github.com/madflojo/tarmac/callbacks/logging"
-	"github.com/madflojo/tarmac/callbacks/metrics"
-	sqlstore "github.com/madflojo/tarmac/callbacks/sql"
-	"github.com/madflojo/tarmac/wasm"
+	"github.com/madflojo/tarmac/pkg/callbacks"
+	"github.com/madflojo/tarmac/pkg/callbacks/httpclient"
+	"github.com/madflojo/tarmac/pkg/callbacks/kvstore"
+	"github.com/madflojo/tarmac/pkg/callbacks/logging"
+	"github.com/madflojo/tarmac/pkg/callbacks/metrics"
+	sqlstore "github.com/madflojo/tarmac/pkg/callbacks/sql"
+	"github.com/madflojo/tarmac/pkg/wasm"
 	"github.com/madflojo/tasks"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/sirupsen/logrus"
