+++ example/toolkit/main.go
@@ -6,8 +6,8 @@ package main
 
 import (
 	// Import Tarmac Callbacks Router and Desired Callback Capabilities
-	"github.com/madflojo/tarmac/callbacks"
-	"github.com/madflojo/tarmac/callbacks/logging"
+	"github.com/madflojo/tarmac/pkg/callbacks"
+	"github.com/madflojo/tarmac/pkg/callbacks/logging"
 	// Import Tarmac WASM Engine
 	"github.com/madflojo/tarmac/wasm"
 	"github.com/sirupsen/logrus"
