+++ pkg/callbacks/httpclient/httpclient.go
@@ -3,8 +3,8 @@ Package httpclient is part of the Tarmac suite of Host Callback packages. This p
 ability to provide WASM functions with a host callback interface that provides HTTP client capabilities.
 
 	import (
-		"github.com/madflojo/tarmac/callbacks"
-		"github.com/madflojo/tarmac/callbacks/httpclient"
+		"github.com/madflojo/tarmac/pkg/callbacks"
+		"github.com/madflojo/tarmac/pkg/callbacks/httpclient"
 	)
 
 	func main() {
