+++ pkg/callbacks/logging/logging.go
@@ -3,8 +3,8 @@ Package logging is part of the Tarmac suite of Host Callback packages. This pack
 to provide WASM functions with a host callback interface that provides logging capabilities.
 
 	import (
-		"github.com/madflojo/tarmac/callbacks"
-		"github.com/madflojo/tarmac/callbacks/logging"
+		"github.com/madflojo/tarmac/pkg/callbacks"
+		"github.com/madflojo/tarmac/pkg/callbacks/logging"
 	)
 
 	func main() {
