+++ pkg/callbacks/metrics/metrics.go
@@ -3,8 +3,8 @@ Package metrics is part of the Tarmac suite of Host Callback packages. This pack
 to provide WASM functions with a host callback interface that provides metrics tracking capabilities.
 
 	import (
-		"github.com/madflojo/tarmac/callbacks"
-		"github.com/madflojo/tarmac/callbacks/metrics"
+		"github.com/madflojo/tarmac/pkg/callbacks"
+		"github.com/madflojo/tarmac/pkg/callbacks/metrics"
 	)
 
 	func main() {
