+++ pkg/wasm/wasm.go
