+++ pkg/wasm/wasm_test.go
